+++ provider/core_route_table_resource.go
@@ -24,9 +24,10 @@ func RouteTableResource() *schema.Resource {
 		Delete:   deleteRouteTable,
 		Schema: map[string]*schema.Schema{
 			"compartment_id": {
-				Type:     schema.TypeString,
-				Required: true,
-				ForceNew: true,
+				Type:             schema.TypeString,
+				Optional:         true,
+				ForceNew:         true,
+				DiffSuppressFunc: crud.DefaultResourceSuppressDiff,
 			},
 			"display_name": {
 				Type:     schema.TypeString,
@@ -36,6 +37,12 @@ func RouteTableResource() *schema.Resource {
 			"id": {
 				Type:     schema.TypeString,
 				Computed: true,
+			},
+			"default_id": {
+				Type:          schema.TypeString,
+				Optional:      true,
+				ConflictsWith: []string{"compartment_id", "vcn_id"},
+				ForceNew:      true,
 			},
 			"route_rules": {
 				Type:     schema.TypeList,
@@ -66,9 +73,10 @@ func RouteTableResource() *schema.Resource {
 				Computed: true,
 			},
 			"vcn_id": {
-				Type:     schema.TypeString,
-				Required: true,
-				ForceNew: true,
+				Type:             schema.TypeString,
+				Optional:         true,
+				ForceNew:         true,
+				DiffSuppressFunc: crud.DefaultResourceSuppressDiff,
 			},
 		},
 	}
@@ -136,6 +144,14 @@ func (s *RouteTableResourceCrud) State() string {
 }
 
 func (s *RouteTableResourceCrud) Create() (e error) {
+	// If we are creating a default resource, then don't have to
+	// actually create it. Just set the ID and update it.
+	if defaultId := s.D.Get("default_id").(string); defaultId != "" {
+		s.D.SetId(defaultId)
+		e = s.Update()
+		return
+	}
+
 	compartmentID := s.D.Get("compartment_id").(string)
 	vcnID := s.D.Get("vcn_id").(string)
 
@@ -147,7 +163,6 @@ func (s *RouteTableResourceCrud) Create() (e error) {
 	if e != nil {
 		return e
 	}
-
 	s.Res, e = s.Client.CreateRouteTable(compartmentID, vcnID, rr, opts)
 
 	return
@@ -157,6 +172,15 @@ func (s *RouteTableResourceCrud) Get() (e error) {
 	res, e := s.Client.GetRouteTable(s.D.Id())
 	if e == nil {
 		s.Res = res
+
+		// If this is a default resource that we removed earlier, then
+		// we need to assume that the parent resource will remove it
+		// and notify terraform of it. Otherwise, terraform will
+		// see that the resource is still available and error out
+		if s.D.Get("default_id") != "" &&
+			s.D.Get("state") == baremetal.ResourceTerminated {
+			s.Res.State = baremetal.ResourceTerminated
+		}
 	}
 	return
 }
@@ -198,6 +222,15 @@ func (s *RouteTableResourceCrud) SetData() {
 }
 
 func (s *RouteTableResourceCrud) Delete() (e error) {
+	if s.D.Get("default_id") != "" {
+		// We can't actually delete a default resource.
+		// Instead, remove all the route rules first and mark it as deleted.
+		s.D.Set("route_rules", nil)
+		e = s.Update()
+		s.D.Set("state", baremetal.ResourceTerminated)
+		return
+	}
+
 	return s.Client.DeleteRouteTable(s.D.Id(), nil)
 }
